backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
arch==7.2.0
backports.asyncio.runner==1.2.0
certifi==2025.8.3
click==8.1.8
diskcache==5.6.3
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
hmmlearn==0.3.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
llama_cpp_python==0.3.16
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numba==0.60.0
numpy==2.0.2
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pluggy==1.6.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==22.3
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
rich==14.1.0
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
statsmodels==0.14.5
stumpy==1.13.0
threadpoolctl==3.6.0
tomli==2.2.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
